GPU: Corrections and feedback to GPU Clock.
The GPU runs at different clockspeeds and paces than the CPU. It 
requires its own timer in order to keep it at check. Sadly a cycle timer 
is unfeasible for the GPU since it's pretty hard to accurately estimate 
the length of time actions like Draw, DispatchCompute and Clear can take 
(varying very unproportionally)
Using the CPU's clock also has it's disadvantages:

- If the CPU is idle while the GPU is running, the timer will advance 
unproportionally due to how CPU's idle timing works.

- If the GPU is synced, the timer won't advance until control is given 
back to the CPU.

For all these reasons, it has been decided to use a host timer for the 
GPU.