Unicorn: Disable Unicorn Memory Mapping through a setting.
Unicorn maps it's own memory for running. Unicorn's mapping is slow, 
buggy and produces memory leaks. However, dynarmic have implemented all 
the memory instructions in ARMv8. With this PR we disable it through a 
setting and kept in for developers to test ARMv7A if not all the 
instructions are implemented.