CMakeLists: Use -Wno-sign-conversion instead of -Wno-error=sign-conversion
The preceding -Werror=conversion implicitly enables -Wsign-conversion,
but only on Clang, not GCC.  So on GCC, -Wno-error=sign-conversion would
do nothing (absent custom flags supplied on the command line) because
-Wsign-conversion wasn't enabled in the first place.  On Clang, it would
prevent the build from erroring out, but you're still left with a ton of
warnings.

Perhaps the codebase ought to be -Wsign-conversion clean, but it's
currently very far from it, and as long as more-typical GCC users aren't
seeing the warning, Clang users shouldn't either.