GPU: Prequeue frame swaps.
The idea of this change is to fix an issue that has occured since I
introduced the fence manager. Originally frames were sent exactly when
the fence was met. However, now fence release is delayed until the host
gpu catches up, this causes swaps to be delayed further than they
should.

This commit aims to solve that issue by prequeuing frames.

This solves the bad frame pacing on MHR.