config: Make asynchronous shaders OpenGL only
Since the release of Project Hades, the shader de/recompiler rewrite, shader compilation under the Vulkan API has improved significantly compared to the previous shader recompiler.
This setting was conceived when shader compilation was significantly slower than it is now, and thus, its utility is greatly diminished on the Vulkan backend.
This makes the setting exclusive to OpenGL, where its still provides a significant benefit.