remove an error from end of Visual Studio build process
I built yuzu with MSVC today (Visual Studio 2022 hype?), and at the very bottom there was a message stating

27>'/NJH' is not recognized as an internal or external command,
27>operable program or batch file.

The cause of this is a newline at the end of the requirements.txt in the precompiled ffmpeg dlls

Flags intended to be sent to robocopy are instead attempted to be executed as a command.

As this is normal, I believe the build scripts need to adjust to fit what is provided.

externals/cmake-modules/WindowsCopyFiles.cmake has the actual invocation of robocopy

The precompiled dlls are from here, and will be in build/externals
https://github.com/yuzu-emu/ext-windows-bin/tree/master/ffmpeg