Super Cool Workaround for SSL ca-certificate issue
usr/lib/libssl.so.1.1 in various linux distributions have differing
opinions on where to locate the bundle of certificates used to properly
handle https traffic.

Ubuntu 18.04: /etc/ssl/certs/ca-certificates.crt
Manjaro (Arch): /etc/ssl/cert.pem

In the AppImage libssl.so.1.1 is being shipped.

Notable because we build the AppImage on a Ubuntu base, and the steam
decks use arch.

This is a hastily thrown together PR, so should not be merged without
feedback

The AppRun shell script is modified to specify SSL_CERT_FILE in
environment