Make configure dialog non-blocking
Currently, when the configure dialog is open, it prevents inputs to the main
window. This is annoying if you are trying to change your keybinds as you have
to constantly reopen the configure dialog to test your keys.

To make the window non-blocking, we use .show() instead of .exec() on the
configure dialog, and have a slot handle when the confirmation buttons (ok,
cancel, x) are pressed.

However, entering the configuration dialog disables inputs to prevent sending
inputs to the running game while changing inputs. We modify this by preventing
inputs from being sent to the game only when focus is not on the main menu. The
steps are as follows:
- Determine focus in/focus out on main window with QGuiApplication::focusWindowChanged
- Send focus in/focus out signal to ConfigureDialog slot
- ConfigureDialog slot calls EnableConfiguration/DisableConfiguration for ConfigureInput
- ConfigureInput calls EnableConfiguration/DisableConfiguration for ConfigureInputPlayer

Added:
- Signal FocusIn/FocusOut for GMainWindow
- EnableConfiguration()/DisableConfiguration() for ConfigureInputPlayer and
  ConfigureInput to control whether inputs to the game should be blocked.
  ConfigureInput::EnableConfiguration/DisableConfiguration simply forwards the
  call to ConfigureInputPlayer.