Check that the Extent3D TileShift only operates on a 3D ASTC Texture by checking Mip count
It appears that previously the Extent3D Tileshift could apply to Mip 0 of 2D Textures in some cases, this is responsible for breaking Asterix and Obelix, the Recall animation in TOTK and an older version of the Tanslation pack for Tsukihime -A Piece of Blue Glass Moon.

I'm open to someone with a better explanation for why these changes fix things, while not regressing TOTKs gloom.