android: InputHandler: Only handle inputs for player 1
Since we don't have controller remapping on android at the moment, we shouldn't be trying to handle each controller slot automatically like this. In some cases android can pick up input devices that a user doesn't want to control yuzu with and we end up forwarding their actual controller inputs to a disconnected controller. This will stay hardcoded to the player 1 devices until we can get remapping in order.