settings; Don't make CPU settings dependant on cpu debugging flag
When CPU debugging is disabled, none of the Debug options in the CPU tab get applied without a clear indication. The CPU debugging option is not even in the same tab.
This is incredibly bad UX and lead to me believing I had fastmem disabled, while in fact it was still on.